+++ src/leapfrog_integrator.cpp
@@ -8,7 +8,7 @@ void leapfrog_integrator::forward_half_step_p(double a) {
     g->compute_potential(a);
 
 #pragma omp parallel for
-    for (int i = 0; i < g->plist->num_particles; ++i) {
+    for (long i = 0; i < g->plist->num_particles; ++i) {
         g->plist->p->index(i, 0) += g->cosmo.f(a) * g->particle_accel(i, 0) * delta_a / 2;
         g->plist->p->index(i, 1) += g->cosmo.f(a) * g->particle_accel(i, 1) * delta_a / 2;
         g->plist->p->index(i, 2) += g->cosmo.f(a) * g->particle_accel(i, 2) * delta_a / 2;
@@ -19,7 +19,7 @@ void leapfrog_integrator::backward_half_step_p(double a) {
     g->compute_potential(a);
 
 #pragma omp parallel for
-    for (int i = 0; i < g->plist->num_particles; ++i) {
+    for (long i = 0; i < g->plist->num_particles; ++i) {
         g->plist->p->index(i, 0) -= g->cosmo.f(a) * g->particle_accel(i, 0) * delta_a / 2;
         g->plist->p->index(i, 1) -= g->cosmo.f(a) * g->particle_accel(i, 1) * delta_a / 2;
         g->plist->p->index(i, 2) -= g->cosmo.f(a) * g->particle_accel(i, 2) * delta_a / 2;
@@ -30,7 +30,7 @@ void leapfrog_integrator::full_step(double a) {
     g->compute_potential(a);
 
 #pragma omp parallel for
-    for (int i = 0; i < g->plist->num_particles; ++i) {
+    for (long i = 0; i < g->plist->num_particles; ++i) {
         g->plist->p->index(i, 0) += g->cosmo.f(a) * g->particle_accel(i, 0) * delta_a;
         g->plist->p->index(i, 1) += g->cosmo.f(a) * g->particle_accel(i, 1) * delta_a;
         g->plist->p->index(i, 2) += g->cosmo.f(a) * g->particle_accel(i, 2) * delta_a;
@@ -38,7 +38,7 @@ void leapfrog_integrator::full_step(double a) {
 
     a += 0.5 * delta_a;
 #pragma omp parallel for
-    for (int i = 0; i < g->plist->num_particles; ++i) {
+    for (long i = 0; i < g->plist->num_particles; ++i) {
         g->plist->x->index(i, 0) += g->cosmo.f(a) * g->plist->p->index(i, 0) * delta_a / (a * a);
         g->plist->x->index(i, 1) += g->cosmo.f(a) * g->plist->p->index(i, 1) * delta_a / (a * a);
         g->plist->x->index(i, 2) += g->cosmo.f(a) * g->plist->p->index(i, 2) * delta_a / (a * a);
@@ -58,7 +58,7 @@ void leapfrog_integrator::run_sim() {
 
     backward_half_step_p(a);
 
-    for (int i = 0; i < num_steps; ++i) {
+    for (long i = 0; i < num_steps; ++i) {
         full_step(a);
 
         a += delta_a;
