+++ test/test_leapfrog_integrator.cpp
@@ -9,10 +9,10 @@
 #include <fstream>
 
 TEST(leapfrog_integrator, motion_2body_x) {
-    int n = 10;
+    long n = 10;
 
-    int num_particles = 2;
-    int num_dims = 3;
+    long num_particles = 2;
+    long num_dims = 3;
 
     array_2d<double> x(num_particles, num_dims);
     array_2d<double> p(num_particles, num_dims);
@@ -26,7 +26,7 @@ TEST(leapfrog_integrator, motion_2body_x) {
 
     double a0 = 0.5;
     double af = 1;
-    int num_steps = 100;
+    long num_steps = 100;
     double delta_a = (af - a0) / num_steps;
 
     leapfrog_integrator li(a0, af, delta_a, grid);
@@ -42,10 +42,10 @@ TEST(leapfrog_integrator, motion_2body_x) {
 }
 
 TEST(leapfrog_integrator, motion_2body_y) {
-    int n = 10;
+    long n = 10;
 
-    int num_particles = 2;
-    int num_dims = 3;
+    long num_particles = 2;
+    long num_dims = 3;
 
     array_2d<double> x(num_particles, num_dims);
     array_2d<double> p(num_particles, num_dims);
@@ -59,7 +59,7 @@ TEST(leapfrog_integrator, motion_2body_y) {
 
     double a0 = 0.5;
     double af = 1;
-    int num_steps = 100;
+    long num_steps = 100;
     double delta_a = (af - a0) / num_steps;
 
     leapfrog_integrator li(a0, af, delta_a, grid);
@@ -75,10 +75,10 @@ TEST(leapfrog_integrator, motion_2body_y) {
 }
 
 TEST(leapfrog_integrator, motion_2body_z) {
-    int n = 10;
+    long n = 10;
 
-    int num_particles = 2;
-    int num_dims = 3;
+    long num_particles = 2;
+    long num_dims = 3;
 
     array_2d<double> x(num_particles, num_dims);
     array_2d<double> p(num_particles, num_dims);
@@ -92,7 +92,7 @@ TEST(leapfrog_integrator, motion_2body_z) {
 
     double a0 = 0.5;
     double af = 1;
-    int num_steps = 100;
+    long num_steps = 100;
     double delta_a = (af - a0) / num_steps;
 
     leapfrog_integrator li(a0, af, delta_a, grid);
@@ -108,18 +108,18 @@ TEST(leapfrog_integrator, motion_2body_z) {
 }
 
 TEST(leapfrog_integrator, flat_density) {
-    int n = 10;
+    long n = 10;
 
-    int num_particles = n * n * n;
-    int num_dims = 3;
+    long num_particles = n * n * n;
+    long num_dims = 3;
 
     array_2d<double> x(num_particles, num_dims);
     array_2d<double> p(num_particles, num_dims);
 
-    int cur_particle = 0;
-    for (int i = 0; i < n; ++i) {
-        for (int j = 0; j < n; ++j) {
-            for (int k = 0; k < n; ++k) {
+    long cur_particle = 0;
+    for (long i = 0; i < n; ++i) {
+        for (long j = 0; j < n; ++j) {
+            for (long k = 0; k < n; ++k) {
                 x(cur_particle, 0) = i;
                 x(cur_particle, 1) = j;
                 x(cur_particle, 2) = k;
@@ -134,7 +134,7 @@ TEST(leapfrog_integrator, flat_density) {
 
     double a0 = 0.5;
     double af = 1;
-    int num_steps = 100;
+    long num_steps = 100;
     double delta_a = (af - a0) / num_steps;
 
     leapfrog_integrator li(a0, af, delta_a, grid);
@@ -142,9 +142,9 @@ TEST(leapfrog_integrator, flat_density) {
     li.run_sim();
 
     cur_particle = 0;
-    for (int i = 0; i < n; ++i) {
-        for (int j = 0; j < n; ++j) {
-            for (int k = 0; k < n; ++k) {
+    for (long i = 0; i < n; ++i) {
+        for (long j = 0; j < n; ++j) {
+            for (long k = 0; k < n; ++k) {
                 EXPECT_NEAR(i, x(cur_particle, 0), 1e-8);
                 EXPECT_NEAR(j, x(cur_particle, 1), 1e-8);
                 EXPECT_NEAR(k, x(cur_particle, 2), 1e-8);
@@ -156,18 +156,18 @@ TEST(leapfrog_integrator, flat_density) {
 }
 
 TEST(leapfrog_integrator, alternating_density) {
-    int n = 10;
+    long n = 10;
 
-    int num_particles = n * n * n;
-    int num_dims = 3;
+    long num_particles = n * n * n;
+    long num_dims = 3;
 
     array_2d<double> x(num_particles, num_dims);
     array_2d<double> p(num_particles, num_dims);
 
-    int cur_particle = 0;
-    for (int i = 0; i < n; ++i) {
-        for (int j = 0; j < n; ++j) {
-            for (int k = 0; k < n; ++k) {
+    long cur_particle = 0;
+    for (long i = 0; i < n; ++i) {
+        for (long j = 0; j < n; ++j) {
+            for (long k = 0; k < n; ++k) {
                 x(cur_particle, 0) = i - i % 2;
                 x(cur_particle, 1) = j - j % 2;
                 x(cur_particle, 2) = k - k % 2;
@@ -182,7 +182,7 @@ TEST(leapfrog_integrator, alternating_density) {
 
     double a0 = 0.5;
     double af = 1;
-    int num_steps = 100;
+    long num_steps = 100;
     double delta_a = (af - a0) / num_steps;
 
     leapfrog_integrator li(a0, af, delta_a, grid);
@@ -190,9 +190,9 @@ TEST(leapfrog_integrator, alternating_density) {
     li.run_sim();
 
     cur_particle = 0;
-    for (int i = 0; i < n; ++i) {
-        for (int j = 0; j < n; ++j) {
-            for (int k = 0; k < n; ++k) {
+    for (long i = 0; i < n; ++i) {
+        for (long j = 0; j < n; ++j) {
+            for (long k = 0; k < n; ++k) {
                 EXPECT_NEAR(i - i % 2, x(cur_particle, 0), 1e-8);
                 EXPECT_NEAR(j - j % 2, x(cur_particle, 1), 1e-8);
                 EXPECT_NEAR(k - k % 2, x(cur_particle, 2), 1e-8);
@@ -204,18 +204,18 @@ TEST(leapfrog_integrator, alternating_density) {
 }
 
 TEST(leapfrog_integrator, particle_cube) {
-    int n = 10;
+    long n = 10;
 
-    int num_particles = 8;
-    int num_dims = 3;
+    long num_particles = 8;
+    long num_dims = 3;
 
     array_2d<double> x(num_particles, num_dims);
     array_2d<double> p(num_particles, num_dims);
 
-    int cur_particle = 0;
-    for (int i = 0; i < 2; ++i) {
-        for (int j = 0; j < 2; ++j) {
-            for (int k = 0; k < 2; ++k) {
+    long cur_particle = 0;
+    for (long i = 0; i < 2; ++i) {
+        for (long j = 0; j < 2; ++j) {
+            for (long k = 0; k < 2; ++k) {
                 x(cur_particle, 0) = 2 * i;
                 x(cur_particle, 1) = 2 * j;
                 x(cur_particle, 2) = 2 * k;
@@ -232,7 +232,7 @@ TEST(leapfrog_integrator, particle_cube) {
 
     double a0 = 0.5;
     double af = 1;
-    int num_steps = 100;
+    long num_steps = 100;
     double delta_a = (af - a0) / num_steps;
 
     leapfrog_integrator li(a0, af, delta_a, grid);
@@ -240,9 +240,9 @@ TEST(leapfrog_integrator, particle_cube) {
     li.run_sim();
 
     cur_particle = 0;
-    for (int i = 0; i < 2; ++i) {
-        for (int j = 0; j < 2; ++j) {
-            for (int k = 0; k < 2; ++k) {
+    for (long i = 0; i < 2; ++i) {
+        for (long j = 0; j < 2; ++j) {
+            for (long k = 0; k < 2; ++k) {
                 EXPECT_LT(fabs(x(cur_particle, 0) - 1), 1);
                 EXPECT_LT(fabs(x(cur_particle, 1) - 1), 1);
                 EXPECT_LT(fabs(x(cur_particle, 2) - 1), 1);
@@ -254,18 +254,18 @@ TEST(leapfrog_integrator, particle_cube) {
 }
 
 TEST(leapfrog_integrator, pos_write) {
-    int n = 10;
+    long n = 10;
 
-    int num_particles = 8;
-    int num_dims = 3;
+    long num_particles = 8;
+    long num_dims = 3;
 
     array_2d<double> x(num_particles, num_dims);
     array_2d<double> p(num_particles, num_dims);
 
-    int cur_particle = 0;
-    for (int i = 0; i < 2; ++i) {
-        for (int j = 0; j < 2; ++j) {
-            for (int k = 0; k < 2; ++k) {
+    long cur_particle = 0;
+    for (long i = 0; i < 2; ++i) {
+        for (long j = 0; j < 2; ++j) {
+            for (long k = 0; k < 2; ++k) {
                 x(cur_particle, 0) = 2 * i;
                 x(cur_particle, 1) = 2 * j;
                 x(cur_particle, 2) = 2 * k;
@@ -282,7 +282,7 @@ TEST(leapfrog_integrator, pos_write) {
 
     double a0 = 0.5;
     double af = 1;
-    int num_steps = 100;
+    long num_steps = 100;
     double delta_a = (af - a0) / num_steps;
 
     std::filesystem::create_directory("test_output");
@@ -305,7 +305,7 @@ TEST(leapfrog_integrator, pos_write) {
     array_2d<double> x_written(num_particles, num_dims);
     file.read((char *) x_written.data, sizeof(double) * num_particles * num_dims);
 
-    for (int i = 0; i < num_particles; ++i) {
+    for (long i = 0; i < num_particles; ++i) {
         EXPECT_DOUBLE_EQ(x(i, 0), x_written(i, 0));
         EXPECT_DOUBLE_EQ(x(i, 1), x_written(i, 1));
         EXPECT_DOUBLE_EQ(x(i, 2), x_written(i, 2));
@@ -317,18 +317,18 @@ TEST(leapfrog_integrator, pos_write) {
 }
 
 TEST(leapfrog_integrator, mom_write) {
-    int n = 10;
+    long n = 10;
 
-    int num_particles = 8;
-    int num_dims = 3;
+    long num_particles = 8;
+    long num_dims = 3;
 
     array_2d<double> x(num_particles, num_dims);
     array_2d<double> p(num_particles, num_dims);
 
-    int cur_particle = 0;
-    for (int i = 0; i < 2; ++i) {
-        for (int j = 0; j < 2; ++j) {
-            for (int k = 0; k < 2; ++k) {
+    long cur_particle = 0;
+    for (long i = 0; i < 2; ++i) {
+        for (long j = 0; j < 2; ++j) {
+            for (long k = 0; k < 2; ++k) {
                 x(cur_particle, 0) = 2 * i;
                 x(cur_particle, 1) = 2 * j;
                 x(cur_particle, 2) = 2 * k;
@@ -345,7 +345,7 @@ TEST(leapfrog_integrator, mom_write) {
 
     double a0 = 0.5;
     double af = 1;
-    int num_steps = 100;
+    long num_steps = 100;
     double delta_a = (af - a0) / num_steps;
 
     std::filesystem::create_directory("test_output");
@@ -368,7 +368,7 @@ TEST(leapfrog_integrator, mom_write) {
     array_2d<double> p_written(num_particles, num_dims);
     file.read((char *) p_written.data, sizeof(double) * num_particles * num_dims);
 
-    for (int i = 0; i < num_particles; ++i) {
+    for (long i = 0; i < num_particles; ++i) {
         EXPECT_DOUBLE_EQ(p(i, 0), p_written(i, 0));
         EXPECT_DOUBLE_EQ(p(i, 1), p_written(i, 1));
         EXPECT_DOUBLE_EQ(p(i, 2), p_written(i, 2));
